% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purpose}
\alias{purpose}
\title{Purpose in Life tScores and Percentiles}
\format{
A data frame with 52 rows and 3 variables
\describe{
\item{raw}{purpose in life mean score}
\item{percents}{percentile for given mean score}
\item{tScores}{tScore for given mean score}
}
}
\usage{
purpose
}
\description{
A dataset containing the tScores and Percentiles for population norms of purpose in life
}
\keyword{datasets}
