% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_purpose.R
\name{score_purpose}
\alias{score_purpose}
\title{Score Purpose in Life Questionnaire}
\usage{
score_purpose(input_df, form = "full", missing_threshold = 0.5, tscore = F)
}
\arguments{
\item{input_df}{data frame}

\item{form}{string}

\item{missing_threshold}{double}

\item{tscore}{logical}
}
\value{
data frame
}
\description{
Calculate mean scores, tScores, and percentiles from raw Purpose in Life data. Input raw data from survey as a data frame, and receive a scored data frame as output
}
\examples{
score_purpose(test_data)
}
