% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data of the full Purpose in Life Scale}
\format{
A data frame with 10 rows and 13 variables
\describe{
\item{SubjectID}{dummy subject ids}
\item{purpose_n}{Purpose in Life Item randomized value}
}
}
\usage{
test_data
}
\description{
A dataset containing the test data for the purpose in life scale including example missing data
}
\keyword{datasets}
